imp: moved tests to same file
